fix: fixed processing account total balances with old omnipool lps
