+++ indexers/liquidity-pools/.env.hydration.local
@@ -55,16 +55,13 @@ SUBSCAN_PRO_API_SECRET=
 SUBSCAN_PROXY_API_CORS_ALLOWED_SUFFIXES=hydration.net::-hydra-app.netlify.app
 SUBSCAN_PROXY_API_CORS_ALLOW_LOCALHOST=true
 
-#PROCESS_FROM_BLOCK=9559980
-#PROCESS_TO_BLOCK=9560090
 
-#PROCESS_FROM_BLOCK=9941810
-PROCESS_FROM_BLOCK=10280130
-PROCESS_TO_BLOCK=10280345
+PROCESS_FROM_BLOCK=10348850
+PROCESS_TO_BLOCK=10349095
 
+#PROCESS_FROM_BLOCK=10350207
+#PROCESS_TO_BLOCK=10350207
 
-#PROCESS_FROM_BLOCK=9942317
-#PROCESS_TO_BLOCK=9942317
 
 
 USE_STORAGE_DICTIONARY=true
